git-svn-id: http://vps6982.inmotionhosting.com/svn/lasi@617 49725315-0998-4498-bb4a-373bc1a52cb2

+++ LASI_Algorithm/Thesuari/Thesaurus.cs
@@ -177,7 +177,7 @@ namespace LASI.Algorithm.Thesauri
                     }
                 }
 
-                return (float)similarCount / (float)(inner.Words.GetNouns().Count() * (float)outer.Words.GetNouns().Count());
+                return similarCount / (inner.Words.GetNouns().Count() * outer.Words.GetNouns().Count());
             }
             else
                 return 1;
